(bug) remove untestable notification addon, womp womp
turns out ember-notify isnt really testable in acceptance tests
unless you stub out the components and services, which doesn't make
a ton of sense because that would mean your tests on verify that a
method gets called, rather than testing if the real app works as expected.
there's an open issue for this on github but it's a year old.